+++ common/src/main/java/net/berserker_rpg/config/TweaksConfig.java
@@ -3,6 +3,6 @@ package net.berserker_rpg.config;
 public class TweaksConfig {
     public boolean ignore_items_required_mods = false;
     public float bloody_strike_self_damage = 0.7F;
-    public float blood_reckoning_absoprtion_to_heal = 0.4F;
-    public float blood_reckoning_base_heal = 4.0F;
+    public float blood_reckoning_missing_health_to_absorption = 0.5F;
+    public float blood_reckoning_absorption_to_heal_on_expire = 0.5F;
 }