+++ common/src/main/java/net/berserker_rpg/effect/BerserkerEffects.java
@@ -14,6 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static net.berserker_rpg.BerserkerClassMod.MOD_ID;
+import static net.berserker_rpg.BerserkerClassMod.tweaksConfig;
 
 public class BerserkerEffects {
     public static final List<Effects.Entry> entries = new ArrayList<>();
@@ -53,17 +54,25 @@ public class BerserkerEffects {
                     )
             ))
     ));
+    public static final Effects.Entry BLOOD_RECKONING = add(new Effects.Entry(
+            Identifier.of(MOD_ID, "blood_reckoning"),
+            "Blood Reckoning",
+            "Grants absorption for missing health. Heals for a portion of your remaining absorption when the effect runs out.",
+            new BloodReckoningEffect(StatusEffectCategory.BENEFICIAL, 0xf70000),
+            new EffectConfig(List.of(
+                    new AttributeModifier(
+                            EntityAttributes.GENERIC_MAX_ABSORPTION.getIdAsString(),
+                            20.0F,
+                            EntityAttributeModifier.Operation.ADD_VALUE
+                    )
+            ))
+    ));
     public static final Effects.Entry OUTRAGE = add(new Effects.Entry(
             Identifier.of(MOD_ID, "outrage"),
             "Outrage",
-            "Clears harmful status effects when applied, increases attack speed & damage.",
+            "Clears harmful status effects when applied, increases attack damage.",
             new OutrageEffect(StatusEffectCategory.BENEFICIAL, 0xf70000),
             new EffectConfig(List.of(
-                    new AttributeModifier(
-                            EntityAttributes.GENERIC_ATTACK_SPEED.getIdAsString(),
-                            0.1F,
-                            EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE
-                    ),
                     new AttributeModifier(
                             EntityAttributes.GENERIC_ATTACK_DAMAGE.getIdAsString(),
                             0.1F,
@@ -71,8 +80,27 @@ public class BerserkerEffects {
                     )
             ))
     ));
+    public static final Effects.Entry APPREHEND = add(new Effects.Entry(
+            Identifier.of(MOD_ID, "apprehend"),
+            "Apprehend",
+            "Lowers armor.",
+            new CustomStatusEffect(StatusEffectCategory.HARMFUL, 0x8a1f1f),
+            new EffectConfig(List.of(
+                    new AttributeModifier(
+                            EntityAttributes.GENERIC_ARMOR.getIdAsString(),
+                            -4.0F,
+                            EntityAttributeModifier.Operation.ADD_VALUE
+                    )
+            ))
+    ));
 
     public static void register(ConfigFile.Effects config) {
+        OnRemoval.configure(BLOOD_RECKONING.effect, (context) -> {
+            var entity = context.entity();
+            float heal = entity.getAbsorptionAmount() * tweaksConfig.value.blood_reckoning_absorption_to_heal_on_expire;
+            entity.heal(heal);
+        });
+
         for (var entry: entries) {
             Synchronized.configure(entry.effect, true);
         }