+++ common/src/main/java/net/berserker_rpg/item/weapons/WeaponsRegister.java
@@ -103,11 +103,6 @@ public class WeaponsRegister {
     private static final String LNE = "loot_n_explore";
     private static final float lneWeaponSpellPower = 2.0F;
     private static final float lneAxeAttackDamage = 15.0F;
-    ///PASSIVE SPELLS
-    public static Identifier dragonclaw = Identifier.of("loot_n_explore", "dragonclaw");
-    public static Identifier avalanche = Identifier.of("loot_n_explore", "avalanche");
-    public static Identifier waterbomb = Identifier.of("loot_n_explore", "waterbomb");
-    public static Identifier wither_pulse = Identifier.of("loot_n_explore", "wither_pulse");
     //Registration
     public static void register(Map<String, WeaponConfig> configs) {
         if(FabricLoader.getInstance().isModLoaded(BETTER_NETHER) || BerserkerClassMod.tweaksConfig.value.ignore_items_required_mods){
@@ -140,28 +135,28 @@ public class WeaponsRegister {
                     .translatedName("Dragons Conquest")
                     .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
                     .attribute(AttributeModifier.bonus(SpellSchools.ARCANE.id, lneWeaponSpellPower))
-                    .spellContainer(SpellContainers.forMagicWeapon().withSpellId(dragonclaw))
+                    .withAdditionalSpell("loot_n_explore:dragonclaw")
                     .rarity = Rarity.RARE;
             berserker_axes( "elder_guardian_berserker_axe",
                     Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.PRISMARINE_SHARD)),lneAxeAttackDamage)
                     .translatedName("Sunken Captain")
                     .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
                     .attribute(AttributeModifier.bonus(MoreSpellSchools.WATER.id, lneWeaponSpellPower))
-                    .spellContainer(SpellContainers.forMagicWeapon().withSpellId(waterbomb))
+                    .withAdditionalSpell("loot_n_explore:waterbomb")
                     .rarity = Rarity.RARE;
             berserker_axes( "wither_berserker_axe",
                     Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.BONE)),lneAxeAttackDamage)
                     .translatedName("Soul Ripper")
                     .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
                     .attribute(AttributeModifier.bonus(SpellSchools.SOUL.id, lneWeaponSpellPower))
-                    .spellContainer(SpellContainers.forMagicWeapon().withSpellId(wither_pulse))
+                    .withAdditionalSpell("loot_n_explore:wither_pulse")
                     .rarity = Rarity.RARE;
             berserker_axes( "glacial_berserker_axe",
                     Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.ICE)),lneAxeAttackDamage)
                     .translatedName("Norse Raid Axe")
                     .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
                     .attribute(AttributeModifier.bonus(SpellSchools.FROST.id, lneWeaponSpellPower))
-                    .spellContainer(SpellContainers.forMagicWeapon().withSpellId(avalanche))
+                    .withAdditionalSpell("loot_n_explore:avalanche")
                     .rarity = Rarity.RARE;
 // ADD THIS IN LNE OR BERSERKER LNE MOD
             /*