+++ common/src/main/java/net/berserker_rpg/spell/CustomSpellImpacts.java
@@ -1,6 +1,6 @@
 package net.berserker_rpg.spell;
 
-import net.berserker_rpg.spell.custom_spell_impacts.BloodReckoningImpact;
+import net.berserker_rpg.spell.custom_spell_impacts.NortherhersGuillotineImpact;
 import net.minecraft.util.Identifier;
 import net.spell_engine.api.spell.event.SpellHandlers;
 
@@ -10,8 +10,8 @@ public class CustomSpellImpacts {
 
     public static void registerCustomImpacts(){
         SpellHandlers.registerCustomImpact(
-                Identifier.of(MOD_ID, "blood_reckoning"),
-                new BloodReckoningImpact()
+                Identifier.of(MOD_ID, "northerners_guillotine"),
+                new NortherhersGuillotineImpact()
         );
     }
 }